+++ sdk/servicebus/azure-messaging-servicebus/src/main/java/com/azure/messaging/servicebus/ServiceBusReceiverClient.java
@@ -399,9 +399,11 @@ public final class ServiceBusReceiverClient implements AutoCloseable {
             .timeout(operationTimeout);
 
         final Flux<ServiceBusReceivedMessage> tracedMessages = tracer.traceSyncReceive("ServiceBus.peekMessages", messages);
-        // Subscribe to message flux so we can kick off this operation, but not to tracing - caller subscriber will take care of it.
-        messages.subscribe();
 
+        // Subscribe to message flux so we can kick off this operation
+        // this works because all messages come in batched messages from management node, which is deserialized as a list and
+        // then returned as Flux.fromIterable.
+        tracedMessages.subscribe();
         return new IterableStream<>(tracedMessages);
     }
 
@@ -450,8 +452,11 @@ public final class ServiceBusReceiverClient implements AutoCloseable {
             sessionId).timeout(operationTimeout);
 
         final Flux<ServiceBusReceivedMessage> tracedMessages = tracer.traceSyncReceive("ServiceBus.peekMessages", messages);
-        // Subscribe to message flux so we can kick off this operation, but not to tracing - caller subscriber will take care of it.
-        messages.subscribe();
+
+        // Subscribe to message flux so we can kick off this operation
+        // this works because all messages come in batched messages from management node, which is deserialized as a list and
+        // then returned as Flux.fromIterable.
+        tracedMessages.subscribe();
 
         return new IterableStream<>(tracedMessages);
     }
@@ -539,8 +544,8 @@ public final class ServiceBusReceiverClient implements AutoCloseable {
         final Flux<ServiceBusReceivedMessage> messagesFlux = emitter.asFlux();
         final Flux<ServiceBusReceivedMessage> tracedMessages = tracer.traceSyncReceive("ServiceBus.receiveMessages", messagesFlux);
 
-        // Subscribe to message flux so we can kick off this operation, but not to tracing - caller subscriber will take care of it.
-        messagesFlux.subscribe();
+        // Subscribe to message flux so we can kick off this operation
+        tracedMessages.subscribe();
 
         return new IterableStream<>(tracedMessages);
     }
@@ -614,8 +619,11 @@ public final class ServiceBusReceiverClient implements AutoCloseable {
             sessionId).timeout(operationTimeout);
 
         final Flux<ServiceBusReceivedMessage> tracedMessages = tracer.traceSyncReceive("ServiceBus.receiveDeferredMessageBatch", messages);
-        // Subscribe so we can kick off this operation.
-        messages.subscribe();
+
+        // Subscribe to message flux so we can kick off this operation
+        // this works because all messages come in batched messages from management node, which is deserialized as a list and
+        // then returned as Flux.fromIterable.
+        tracedMessages.subscribe();
 
         return new IterableStream<>(tracedMessages);
     }